lift import event data function to context
